requirements.txt
# AI Solar 2.0 - Requirements
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# AI Providers
anthropic==0.7.8
openai==1.3.7
requests==2.31.0

# Utilities
pydantic==2.5.1
python-multipart==0.0.6
asyncio==3.4.3
aiohttp==3.9.1
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0